api/requirements.txt
matplotlib
numpy
notebook
tensorflow-addons
tensorflow-model-optimization
tensorflow
flask
pillow
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0

